+++ internal/core/domain/payment.go
@@ -3,9 +3,9 @@ package domain
 import "time"
 
 type Payment struct {
-	CustomerID  *uint  `json:"customerId"`
-	TotalPrice  int    `json:"totalPrice" validate:"required"`
-	PaymentType string `json:"paymentType" validate:"required"`
+	CustomerID  *uint   `json:"customerId"`
+	TotalPrice  float64 `json:"totalPrice" validate:"required"`
+	PaymentType string  `json:"paymentType" validate:"required"`
 }
 
 type PaymentResponse struct {
